Module Graph

module Sig : sig ... end

Signatures for graph implementations.

module Sig_pack : sig ... end

Immediate access to the library: contain a signature gathering an imperative graph signature and all algorithms. Vertices and edges are labeled with integers.

module Dot_ast : sig ... end

AST for DOT file format.

module Util : sig ... end

Some useful operations.

module Blocks : sig ... end

Common implementation to persistent and imperative graphs.

module Persistent : sig ... end

Persistent Graph Implementations.

module Imperative : sig ... end

Imperative Graph Implementations.

module Delaunay : sig ... end

Delaunay triangulation.

module Builder : sig ... end

Graph builders in order to persistent/imperative graphs sharing a same signature.

module Classic : sig ... end

Some classic graphs

module Rand : sig ... end

Random graph generation.

module Oper : sig ... end

Basic operations over graphs

module Components : sig ... end

Strongly connected components.

module Path : sig ... end

Paths

module Cycles : sig ... end

Algorithms related to cycles in directed graphs.

module Nonnegative : sig ... end

Weighted graphs without negative-cycles.

module Traverse : sig ... end

Graph traversal.

module Coloring : sig ... end

k-coloring of undirected graphs.

module Topological : sig ... end

Topological order.

module Kruskal : sig ... end

Kruskal's minimum-spanning-tree algorithm.

module Flow : sig ... end

Algorithms on flows

module Prim : sig ... end

module Dominator : sig ... end

Dominators

module Graphviz : sig ... end

Interface with GraphViz

module Gml : sig ... end

Parser and pretty-printer for GML file format.

module Dot : sig ... end

Parser for DOT file format.

module Pack : sig ... end

Immediate access to the library: provides implementation of imperative graphs labeled with integer as well as algorithms on such graphs.

module Gmap : sig ... end

Graph mapping. Map a graph to another one.

module Minsep : sig ... end

Minimal separators of a graph

module Cliquetree : sig ... end

Construction of the clique tree of a graph and recognition of chordal graphs.

module Mcs_m : sig ... end

Maximal Cardinality Search (MCS-M) algorithm

module Md : sig ... end

Minimum Degree algorithm

module Strat : sig ... end

Strategies

module Fixpoint : sig ... end

Fixpoint computation implemented using the work list algorithm. This module makes writing data-flow analysis easy.

module Leaderlist : sig ... end

The leader list algorithm; it generates a list of basic blocks from a directed graph. A basic block is a forward path of nodes that requires neither branching from nor into.

module Contraction : sig ... end

Edge contraction for directed, edge-labeled graphs

module Graphml : sig ... end

Generic GraphMl Printer

module Merge : sig ... end

Provides functions to extend any module satisfying one of the signatures Sig.P, Sig.I and Builder.S .

module Mincut : sig ... end

Minimal cutset of a graph

module Clique : sig ... end

Graph cliques

module WeakTopological : sig ... end

Weak topological ordering of the vertices of a graph, as described by François Bourdoncle.

module ChaoticIteration : sig ... end

Fixpoint computation with widenings using weak topological orderings as defined by François Bourdoncle and implemented in WeakTopological.

module XDotDraw : sig ... end

Parses xdot drawing operations

module XDot : sig ... end

Reads layout information from xdot ASTs

module DGraphModel : sig ... end

Abstract graph model

module DGraphTreeLayout : sig ... end

module DGraphSubTree : sig ... end

module DGraphTreeModel : sig ... end

module DGraphRandModel : sig ... end