Module Graph.Coloring

k-coloring of undirected graphs.

A k-coloring of a graph g is a mapping c from nodes to \{1,...,k\} such that c(u) <> c(v) for any edge u-v in g.

exception NoColoring

Graph coloring for graphs with integer marks

module type GM = sig ... end

Minimal graph signature for Mark. Sub-signature of Sig.IM.

module Mark (G : GM) : sig ... end

Provide a function for k-coloring a graph with integer marks. The provided function is more efficient that the one provided by functor Make above.

Graph coloring for graphs without marks

module type G = sig ... end

Minimal graph signature for Make. Sub-signature of Sig.G.

module Make (G : G) : sig ... end

Provide a function for k-coloring a graph.