Module Graph.Dominator

Dominators

All of the functions in this module assume that the graph is not modified between calling one of these functions and using the returned functions. Such mutation results in undefined behavior.

author Ivan Jager

exception Unreachable

module type G = sig ... end

module type S = sig ... end

module Make (G : G) : S with type t = G.t and type vertex = G.V.t

module type I = sig ... end

module Make_graph (G : I) : sig ... end