Parameter Parse.L

val node : Dot_ast.node_id -> Dot_ast.attr list -> B.G.V.label

How to build the node label out of the set of attributes

val edge : Dot_ast.attr list -> B.G.E.label

How to build the edge label out of the set of attributes