Module Graph.Leaderlist

The leader list algorithm; it generates a list of basic blocks from a directed graph. A basic block is a forward path of nodes that requires neither branching from nor into.

module type G = sig ... end

Minimal graph signature for leader list algorithm

module Make (G : G) : sig ... end