Module Graph.Minsep

Minimal separators of a graph

Based on the article: Generating all the minimal separators of a graph. by A. Berry, J.-P. Bordat and O.Cogis http://www.isima.fr/berry/generating.html

A set S of vertices is a minimal separator if it exists 2 distinct connected components C and D in G \ S such that each vertex of S has a successor in C and D.

module type G = sig ... end

Minimal signature for computing the minimal separators

module type MINSEP = sig ... end

module P (G : sig ... end) : MINSEP with module G = G

Implementation for a persistent graph

module I (G : sig ... end) : MINSEP with module G = G

Implementation for an imperative graph. Less efficient that the implementation for a persistent graph