Module Imperative.Mark

Mark on vertices. Marks can be used if you want to store some information on vertices: it is more efficient to use marks than an external table.

type graph = t

Type of graphs.

type vertex = vertex

Type of graph vertices.

val clear : graph -> unit

clear g sets all the marks to 0 for all the vertices of g.

val get : vertex -> int

Mark value (in O(1)).

val set : vertex -> int -> unit

Set the mark of the given vertex.