Module Graph.Sig_pack

Immediate access to the library: contain a signature gathering an imperative graph signature and all algorithms. Vertices and edges are labeled with integers.

module type S = sig ... end

Signature gathering an imperative graph signature and all algorithms. Vertices and edges are labeled with integers.