Module WeakTopological.Make

Parameters

module G : G

Signature

val recursive_scc : G.t -> G.V.t -> G.V.t t

recursive_scc g root_g computes a weak topological ordering of the vertices of g, with the general algorithm recursively computing the strongly connected components of g. root_g is taken as the root of the graph and must be a valid vertex of g.