Module Graph__.DGraphTreeModel

module type S = sig ... end

module SubTreeMake : functor (G : Graph.Graphviz.GraphWithDotAttrs) -> sig ... end

This functor creates a model centered on a vertex from a graph

module SubTreeDotModelMake : sig ... end

Creates a model centered on a vertex from a dot model