Parameter Parse.2-L

val node : Graph.Dot_ast.node_id -> Graph.Dot_ast.attr list -> B.G.V.label

How to build the node label out of the set of attributes

val edge : Graph.Dot_ast.attr list -> B.G.E.label

How to build the edge label out of the set of attributes