Module Gmap.Vertex

Provide a mapping function from a mapping of vertices.

Parameters

G_Src : V_SRC

G_Dst : V_DST

Signature

val map : (G_Src.V.t -> G_Dst.vertex) -> G_Src.t -> G_Dst.t

map f g applies f to each vertex of g and so builds a new graph based on g

val filter_map : (G_Src.V.t -> G_Dst.vertex option) -> G_Src.t -> G_Dst.t

filter_map f g applies f to each vertex of g and so builds a new graph based on g; if None is returned by f the vertex is omitted in the new graph.