Parameter Parse.2-L

val node : value_list -> B.G.V.label

How to build the node label out of the set of GML attributes. For example

node [ id 12 label "foo" ]

will call this function with ["id", Int 12; "label", String "foo"]

val edge : value_list -> B.G.E.label

How to build the edge label out of the set of GML attributes