Module Graphviz.NeatoAttributes

type graph = [

| CommonAttributes.graph

| `Margin of float * float
Sets the page margin (included in the page size). Default value is 0.5, 0.5.

| `Start of int
Seed for random number generator.

| `Overlap of bool
Default value is true.

| `Spline of bool
true makes edge splines if nodes don't overlap. Default value is false.

| `Sep of float
Edge spline separation factor from nodes. Default value is 0.0.

]

Attributes of graphs. They include all common graph attributes and several specific ones. All attributes described in the "Neato User's manual, April 10, 2002" are handled.

type vertex = [

| CommonAttributes.vertex

| `Pos of float * float
Initial coordinates of the vertex.

]

Attributes of nodes. They include all common node attributes and several specific ones. All attributes described in the "Neato User's manual, April 10, 2002" are handled.

type edge = [

| CommonAttributes.edge

| `Id of string
Optional value to distinguish multiple edges.

| `Len of float
Preferred length of edge. Default value is 1.0.

| `Weight of float
Strength of edge spring. Default value is 1.0.

]

Attributes of edges. They include all common edge attributes and several specific ones. All attributes described in the "Neato User's manual, April 10, 2002" are handled.

type subgraph = {

sg_name : string;

sg_attributes : vertex list;

sg_parent : string option;

}

Subgraphs have a name and some vertices.