Module Graph__Dominator

Dominators

All of the functions in this module assume that the graph is not modified between calling one of these functions and using the returned functions. Such mutation results in undefined behavior.

author

Ivan Jager