Module Graph__Dot.Parse

Provide a parser for DOT file format.

Parameters

B : Graph.Builder.S

L : sig ... end

Signature

val parse : string -> B.G.t

Parses a dot file

val parse_bounding_box_and_clusters : string -> B.G.t * string * clusters_hash

Parses a dot file and returns the graph, its bounding box and a hash table from clusters to dot attributes