Module Graph__Graphviz.Neato

Parameters

X : sig ... end

Signature

val set_command : string -> unit

Several functions provided by this module run the external program neato. By default, this command is supposed to be in the default path and is invoked by neato. The function set_command allows to set an alternative path at run time.

exception Error of string

val handle_error : ('a -> 'b) -> 'a -> 'b

val fprint_graph : Stdlib.Format.formatter -> X.t -> unit

fprint_graph ppf graph pretty prints the graph graph in the CGL language on the formatter ppf.

val output_graph : Stdlib.out_channel -> X.t -> unit

output_graph oc graph pretty prints the graph graph in the dot language on the channel oc.