Module Graph__Leaderlist.Make

Parameters

G : G

Signature

val leader_lists : G.t -> G.V.t -> G.V.t list list

leader_lists graph root computes the leader lists or basic blocks of the given graph. The node root is always a leader of a basic block.