Module Graph__Minsep

Minimal separators of a graph

Based on the article: Generating all the minimal separators of a graph. by A. Berry, J.-P. Bordat and O.Cogis http://www.isima.fr/berry/generating.html

A set S of vertices is a minimal separator if it exists 2 distinct connected components C and D in G \ S such that each vertex of S has a successor in C and D.