Module Graph__Pack

Immediate access to the library: provides implementation of imperative graphs labeled with integer as well as algorithms on such graphs.

So if you bother functors, you can use this module.