Module Graph__Topological

Topological order.

This functor provides functions which allow iterating over a graph in topological order. Cycles in graphs are allowed. Specification is the following: if vertex x is visited before vertex y then either there is a path from x to y, or there is no path from y to x. In the particular case of a DAG, this simplifies to: if there is an edge from x to y, then x is visited before y.