Module Graph__WeakTopological

Weak topological ordering of the vertices of a graph, as described by François Bourdoncle.

Weak topological ordering is an extension of topological ordering for potentially cyclic graphs.

A hierarchical ordering of a set is a well-parenthesized permutation of its elements with no consecutive (. The elements between two parentheses are called a component, and the first elements of a component is called the head. A weak topological ordering of a graph is a hierarchical ordering of its vertices such that for every edge u -> v of the graph, either u comes (strictly) before v, or v is the head of a component containing u.

One may notice that :

For an acyclic graph, every topological ordering is also a weak topological ordering.

For any graph with the vertices v1, ..., vN, the following trivial weak topological ordering is valid : (v1 (v2 (... (vN))...)).

Weak topological ordering are useful for fixpoint computation (see ChaoticIteration). This module aims at computing weak topological orderings which improve the precision and the convergence speed of these analyses.

author

Thibault Suzanne

see Efficient chaotic iteration strategies with widenings

, François Bourdoncle, Formal Methods in Programming and their Applications, Springer Berlin Heidelberg, 1993