Module Clique.Bron_Kerbosch

Parameters

module G : G

Signature

val maximalcliques : G.t -> G.V.t list list

maximalcliques g computes all the maximal cliques of g using the Bron-Kerbosch algorithm. It returns the sets of vertices belonging to the same maximal clique.