Module Graph.Gmap

Graph mapping. Map a graph to another one.

Mapping of vertices

module type V_SRC = sig ... end

Signature for the source graph.

module type V_DST = sig ... end

Signature for the destination graph.

module Vertex (G_Src : V_SRC) (G_Dst : V_DST) : sig ... end

Provide a mapping function from a mapping of vertices.

Mapping of edges

module type E_SRC = sig ... end

Signature for the source graph.

module type E_DST = sig ... end

Signature for the destination graph.

module Edge (G_Src : E_SRC) (G_Dst : E_DST) : sig ... end

Provide a mapping function from a mapping of edges.