Module Graphviz.CommonAttributes

The CommonAttributes module defines attributes for graphs, vertices and edges that are available in the two engines, dot and neato.

type graph = [

| `Center of bool
Centers the drawing on the page. Default value is false.

| `Fontcolor of color
Sets the font color. Default value is black.

| `Fontname of string
Sets the font family name. Default value is "Times-Roman".

| `Fontsize of int
Sets the type size (in points). Default value is 14.

| `Label of string
Caption for graph drawing.

| `HtmlLabel of string
Caption for graph drawing. In HTML strings, angle brackets must occur in matched pairs, and newlines and other formatting whitespace characters are allowed. In addition, the content must be legal XML, so that the special XML escape sequences for ", &, <, and > may be necessary in order to embed these characters in attribute values or raw text."

| `Orientation of [ `Portrait | `Landscape ]
Sets the page orientation. Default value is `Portrait.

| `Page of float * float
Sets the PostScript pagination unit, e.g 8.5, 11.0.

| `Pagedir of [ `TopToBottom | `LeftToRight ]
Traversal order of pages. Default value is `TopToBottom.

| `Size of float * float
Sets the bounding box of drawing (in inches).

| `OrderingOut
Constrains order of out-edges in a subgraph according to their file sequence

]

Attributes of graphs.

type vertex = [

| `Color of color
Sets the color of the border of the vertex. Default value is black

| `ColorWithTransparency of color_with_transparency
Sets the color of the border of the vertex with a transparency component. Default value is fully opaque black

| `Fontcolor of color
Sets the label font color. Default value is black.

| `Fontname of string
Sets the label font family name. Default value is "Times-Roman".

| `Fontsize of int
Sets the label type size (in points). Default value is 14.

| `Height of float
Sets the minimum height. Default value is 0.5.

| `Label of string
Sets the label printed in the vertex. The string may include escaped newlines \n, \l, or \r for center, left, and right justified lines. Record labels may contain recursive box lists delimited by | .

| `HtmlLabel of string
Like label, in html style. In HTML strings, angle brackets must occur in matched pairs, and newlines and other formatting whitespace characters are allowed. In addition, the content must be legal XML, so that the special XML escape sequences for ", &, <, and > may be necessary in order to embed these characters in attribute values or raw text."

| `Orientation of float
Vertex rotation angle, in degrees. Default value is 0.0.

| `Penwidth of float
Width of the pen (in points) used to draw the border of the node. Default value is 1.0.

| `Peripheries of int
Sets the number of periphery lines drawn around the polygon.

| `Regular of bool
If true, then the polygon is made regular, i.e. symmetric about the x and y axis, otherwise the polygon takes on the aspect ratio of the label. Default value is false.

| `Shape of [ `Ellipse | `Box | `Circle | `Doublecircle | `Diamond | `Plaintext | `Record | `Oval | `Egg | `Triangle | `Invtriangle | `Trapezium | `Invtrapezium | `House | `Invhouse | `Parallelogram | `Doubleoctagon | `Tripleoctagon | `Mdiamond | `Mcircle | `Msquare | `Star | `Underline | `Note | `Tab | `Folder | `Box3d | `Component | `Promoter | `Cds | `Terminator | `Utr | `Primersite | `Restrictionsite | `Fivepoverhang | `Threepoverhang | `Noverhang | `Assembly | `Signature | `Insulator | `Ribosite | `Rnastab | `Proteasesite | `Proteinstab | `Rpromoter | `Rarrow | `Larrow | `Lpromoter | `Polygon of int * float ]
Sets the shape of the vertex. Default value is `Ellipse. `Polygon (i, f) draws a polygon with n sides and a skewing of f.

| `Style of [ `Rounded | `Filled | `Solid | `Dashed | `Dotted | `Bold | `Invis ]
Sets the layout style of the vertex. Several styles may be combined simultaneously.

| `Width of float
Sets the minimum width. Default value is 0.75.

]

Attributes of vertices.

type edge = [

| `Color of color
Sets the edge stroke color. Default value is black.

| `ColorWithTransparency of color_with_transparency
Sets the edge stroke color with a transparency component. Default value is fully opaque black

| `Decorate of bool
If true, draws a line connecting labels with their edges.

| `Dir of [ `Forward | `Back | `Both | `None ]
Sets arrow direction. Default value is `Forward.

| `Fontcolor of color
Sets the label font color. Default value is black.

| `Fontname of string
Sets the label font family name. Default value is "Times-Roman".

| `Fontsize of int
Sets the label type size (in points). Default value is 14.

| `Label of string
Sets the label to be attached to the edge. The string may include escaped newlines \n, \l, or \r for centered, left, or right justified lines.

| `HtmlLabel of string
Like label, in html style. In HTML strings, angle brackets must occur in matched pairs, and newlines and other formatting whitespace characters are allowed. In addition, the content must be legal XML, so that the special XML escape sequences for ", &, <, and > may be necessary in order to embed these characters in attribute values or raw text."

| `Labelfontcolor of color
Sets the font color for head and tail labels. Default value is black.

| `Labelfontname of string
Sets the font family name for head and tail labels. Default value is "Times-Roman".

| `Labelfontsize of int
Sets the font size for head and tail labels (in points). Default value is 14.

| `Penwidth of float
Width of the pen (in points) used to draw the edge. Default value is 1.0.

| `Style of [ `Solid | `Dashed | `Dotted | `Bold | `Invis ]
Sets the layout style of the edge. Several styles may be combined simultaneously.

]

Attributes of edges.