Module Graphviz.Dot

Parameters

module X : sig ... end

Signature

val fprint_graph : Stdlib.Format.formatter -> X.t -> unit

fprint_graph ppf graph pretty prints the graph graph in the CGL language on the formatter ppf.

val output_graph : Stdlib.out_channel -> X.t -> unit

output_graph oc graph pretty prints the graph graph in the dot language on the channel oc.