Module Oper.Neighbourhood

Neighbourhood of vertex / vertices

Parameters

G : sig ... end

Signature

module Vertex_Set : Stdlib.Set.S with type Vertex_Set.elt = G.V.t

val list_from_vertex : G.t -> G.V.t -> G.V.t list

Neighbourhood of a vertex as a list.

val set_from_vertex : G.t -> G.V.t -> Vertex_Set.t

Neighbourhood of a vertex as a set. Less efficient that list_from_vertex.

val list_from_vertices : G.t -> G.V.t list -> G.V.t list

Neighbourhood of a list of vertices as a list.

val set_from_vertices : G.t -> G.V.t list -> Vertex_Set.t

Neighbourhood of a list of vertices as a set. More efficient that list_from_vertices.