Module Graph.Pack

Immediate access to the library: provides implementation of imperative graphs labeled with integer as well as algorithms on such graphs.

So if you bother functors, you can use this module.

module Digraph : Sig_pack.S

Directed imperative graphs with edges and vertices labeled with integer.

module Graph : Sig_pack.S

Undirected imperative graphs with edges and vertices labeled with integer.