Module XDot.Make

Instantiates a module which creates graph layouts from xdot files

Parameters

module G : Graphviz.GraphWithDotAttrs

Signature

module HV : Hashtbl.S with type key = G.V.t

module HE : Map.S with type key = G.E.t

type graph_layout = {

vertex_layouts : node_layout HV.t;

edge_layouts : edge_layout HE.t;

cluster_layouts : (string, cluster_layout) Hashtbl.t;

bbox : bounding_box;

}

Main layout type

exception DotError of string

val layout_of_xdot : xdot_file:string -> G.t -> graph_layout

Extracts a layout of an xdot file

val layout_of_dot : ?cmd:string -> dot_file:string -> G.t -> graph_layout

Using the dot file and graphviz, create an xdot and extracts its layout.